完善姿态检测以及修改抽烟检测对应的接口;

See merge request r_d_center/department-of-ai/Nessus!112

+++ parser/cpp/protocol/refvest.cc
@@ -115,8 +115,8 @@ namespace glasssix::exposing::nessus::Protocol {
 					auto pts = p["key_points"];
 					for(auto j : pts)
 					{
-						key_points.push_back(j["x"].asFloat());
-						key_points.push_back(j["y"].asFloat());
+						key_points.push_back(j["x"].asInt());
+						key_points.push_back(j["y"].asInt());
 						key_points.push_back(j["point_score"].asFloat());
 					}
 					posture.set_key_points(key_points);
